refactor: 💡 extract pattern filter
