+++ Cargo.toml
@@ -149,6 +149,7 @@ typedmap            = "0.5.0"
 urlencoding         = "2.1.3"
 vfs                 = "0.12.0"
 xxhash-rust         = "0.8.10"
+regress            = "0.10.0"
 
 # oxc crates share the same version
 oxc                = { version = "0.24.2", features = ["sourcemap_concurrent", "transformer", "minifier", "semantic", "codegen"] }
