+++ crates/rolldown_common/Cargo.toml
@@ -24,8 +24,6 @@ dashmap            = { workspace = true }
 glob-match         = { workspace = true }
 lightningcss       = { workspace = true }
 oxc                = { workspace = true, features = ["semantic"] }
-regex              = { workspace = true }
-regress            = "0.10.0"
 rolldown_fs        = { workspace = true }
 rolldown_rstr      = { workspace = true }
 rolldown_sourcemap = { workspace = true }
