+++ crates/rolldown_common/src/types/mod.rs
@@ -12,7 +12,6 @@ pub mod exports_kind;
 pub mod external_module_idx;
 pub mod import_record;
 pub mod importer_record;
-pub mod js_regex;
 pub mod member_expr_ref;
 pub mod module_def_format;
 pub mod module_id;
