+++ crates/rolldown_utils/Cargo.toml
@@ -19,6 +19,7 @@ workspace = true
 anyhow      = { workspace = true }
 base64-simd = { workspace = true }
 futures     = { workspace = true }
+glob-match    = { workspace = true }
 indexmap    = { workspace = true, features = ["serde"] }
 infer       = { workspace = true }
 memchr      = { workspace = true }
@@ -26,9 +27,11 @@ mime        = { workspace = true }
 oxc         = { workspace = true }
 phf         = { workspace = true }
 regex       = { workspace = true }
+regress     = { workspace = true }
 rustc-hash  = { workspace = true }
 sugar_path  = { workspace = true }
 xxhash-rust = { workspace = true, features = ["xxh3"] }
+
 
 [target.'cfg(not(target_family = "wasm"))'.dependencies]
 async-scoped = { workspace = true, features = ["use-tokio"] }
