+++ crates/rolldown_utils/src/lib.rs
@@ -17,4 +17,6 @@ pub mod rustc_hash;
 pub mod sanitize_file_name;
 pub mod xxhash;
 pub use bitset::BitSet;
+pub mod js_regex;
+pub mod pattern_filter;
 pub mod unique_arc;
